In #TIMING command
cummulative --> cumulative
cumm --> cumu
Allowed only for stand alone mode.

In #CONSERVATIVECRITERIA command:
reformatted XML to make it work for XmlToTex (lines without < or > are
not removed by XmlToTex even if it is inside and XML expression.
This is an error of the XmlToTex script which should be fixed some day.

In #INNERBOUNDARY command improved the look of ^CFG directive
for the Latex conversion.
This is an error of the XmlToTex script which should be fixed some day.