A serious theoretical mistake has been found and corrected. Strange as it might seem, in partially ionized plasmas (treated as ideal gases), the specific heat (and all the other thermodynamical derivatives) of a mixture is not a weighted average of the specific heats of the mixture components (because of the non-liniear dependence of the ionization degree on temperature and the component fractions).
The only parameter which was calculated erroneously was the specific heat of polymide,
however this error affected all the simulation results with the wall shock.

Since the test results are changed anyway and the reference results should be
changed accordingly, I also removed the redundant condition which limited the range of
parameters for applying the Fermi degeneration.

Backward compatibility: to repeat previous results, you can do the following
manipulations:

1. Uncomment the strange line with the pattern ".and..false." in
util/CRASH/src/ModStatSum.f90 - to disable the fix for thermodynamic derivatives.

2. Set LogGeFermiMin  (equal to -4.0 at the time) to 0.0, to return to the
previously existed restriction on the Fermi-Energy-To-Temperature ratio.


Modified Files:
 	GM/BATSRUS/Param/CRASH/TestOutput/log_hyades2d.log
 	GM/BATSRUS/Param/CRASH/TestOutput/log_levelset.log
 	util/CRASH/doc/Tex/ionization_table5.tex
 	util/CRASH/src/ModFermiGas.f90 util/CRASH/src/ModStatSum.f90


Former-commit-id: 524ee449774167245a498755ec5eeabe7d129db6