Fixed coordinate names for spherical plots (spN and spS)
For unstructured grids replaced the "lookup table" algorithm with
a sorting and averaging algorithm (first implemented and tested by Hongyang).
Now the output is always ordered, which makes test comparisons easier.
The averaging now works for 1D and 2D cuts too, even for large 2D grids.
There is no need to sort and average in the IDL procedures.
If there is a resolution change at the cut plane, the code now keeps
the fine and coarse cell values without averaging.


Former-commit-id: c42e3135541b5e56ec3f4229747980bed1c4b216